Move Card and Deck to new file card.ts
