+++ Server/game.ts
@@ -4,7 +4,7 @@ import { socketTable } from "./users.js";
 import { PlayerInfo } from "./player.js";
 import { Building, BuildingStats } from "./building.js";
 import { Unit, UnitStats } from "./unit.js";
-import { withinRadius } from "./../Client/functions.js";
+import { Card, Deck } from "./card.js";
 import b from "./../Client/buildings.json" assert { type: "json" };
 import u from "./../Client/units.json" assert {type: "json"}; // See if this works or needs parsing
 import c from "./../Client/cards.json" assert {type: "json"};
@@ -230,71 +230,6 @@ class Tile {
     leave() {
         this.occupant = null;
         this.occupantType = null;
-    }
-}
-
-type Card = {
-    name: string, // internal identifier
-    faction: Faction | "N", // Faction this belongs to, or "N" for Neutral
-    cardType: CardType // Building | Unit | Operation
-    cost: number, // cost in money
-    targets: {
-        name: string, // beginning with $
-        type: string, // tile/unit/building/player/card etc.
-        properties: {
-            [key: string]: any
-        }
-    }[]
-    effects: {
-        effect: string // name of the type of effect
-        [key: string]: any
-    }[]
-    /* Sample effect entry:
-    {
-        effect: spawn
-        id: string,
-        loc: string,
-        modifiers: {
-            [key: string]: any
-        }[]
-    }
-    */
-    
-}
-
-class Deck {
-    private cards: Card[] = [];
-    public size = 0; // Is this needed?
-    constructor() {
-        // Maybe add some stuff later
-    }
-    shuffle() {
-        let temp = -1;
-        for (let i = 0; i < this.size - 1; i++) {
-            temp = i + Math.floor((this.size - i) * Math.random());
-            [this.cards[i], this.cards[temp]] = [this.cards[temp], this.cards[i]];
-        }
-    }
-    add(card: Card, quantity=1, copy=false) {
-        if (quantity === 1 && !copy) {
-            this.cards.push(card); // Uses the card directly if quantity is 1 and copy is false
-        } else {
-            for (let i = 0; i < quantity; i++) this.cards.push(deepCopy(card));
-        }
-        this.size += quantity;
-    }
-    draw() {
-        // TODO: Later, decide on a draw algorithm
-        // Temporarily, we'll use each card has twice the chance of being drawn as the previous card except the last two
-        if (this.size === 0) return undefined;
-        let i = 0;
-        let choice = Math.random();
-        while(choice < 0.5 && i < this.size - 1) {
-            choice *= 2;
-            i++;
-        }
-        this.size--;
-        return this.cards.splice(i, 1)[0];
     }
 }
